refactor: extract inline scripts into dedicated source files
Move the two inline scripts from the base layout into their own modules:

- `src/_includes/layouts/_anti-flash.ts`: exports `ANTI_FLASH_SCRIPT`, the
  minified IIFE that reads `localStorage` and sets `data-color-scheme` on
  `<html>` before any CSS renders. The string is still embedded inline in
  `<head>` at build time — an external script loaded with `defer`/`async`
  would arrive too late to suppress the theme flash.

- `src/theme-toggle.page.ts`: a Lume `.page.ts` asset that outputs
  `/theme-toggle.js`. The base layout now references it with
  `<script src="/theme-toggle.js"></script>` instead of inlining the code.

No behaviour changes. `_config.ts` is unchanged: Lume auto-discovers
`.page.ts` files, so the new asset is picked up without additional
`site.add()` calls.
